add InputTypeMixin._random_split and move to data.py
